Add structured file
