+++ src/components/nujovich/styles/main.css
@@ -1,0 +1,23 @@
+.nujovich-advice-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background: #228be2;
+  border-radius: 0.375rem;
+  padding: 1rem;
+  margin: 1rem;
+  box-shadow: 0px 1px 3px 0px rgba(0, 0, 0, 0.1),
+    0px 1px 2px -1px rgba(0, 0, 0, 0.1), 0px 1px 3px 0px rgba(0, 0, 0, 0.1),
+    0px 1px 2px -1px rgba(0, 0, 0, 0.1);
+}
+
+.nujovich-label {
+  padding: 1rem;
+}
+
+.nujovich-button {
+  font-family: "Space Mono", monospace;
+  border-color: #a5d8ff;
+  border-radius: 0.125rem;
+}
